Reorganized the stencil code directory.
