+++ platform/fujitsu-fx100.cmake
@@ -2,6 +2,7 @@
 set(TARGET_SUFFIX               ".cpu")
 
 set(ARCH                        "-KHPC_ACE2")
+set(SIMD_SET                    "HPC_ACE2")
 set(OPENMP_FLAGS                "-Kopenmp")
 set(LAPACK_FLAGS                "-SSL2")
 set(ADDITIONAL_MACRO            "")
