+++ stencil/C/IMCI/current.c
@@ -13,8 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+
+/* Hand-Code Vector processing for Knights Corner */
+
 #include <complex.h>
-#include "./interop.h"
+#include "../interop.h"
 
 extern int NLx, NLy, NLz;
 
@@ -32,8 +35,6 @@ void current_stencil_( double         const* restrict C
                      , double              * restrict G
                      , double              * restrict H
 ) {
-#if defined(__KNC__) || defined(__AVX512F__)
-/* Hand-Code Vector processing for Xeon Phi */
   double complex const* e;
 
   int ix, iy, iz, n;
@@ -190,118 +191,4 @@ void current_stencil_( double         const* restrict C
   *F = _mm512_reduce_add_pd(tt[0]);
   *G = _mm512_reduce_add_pd(tt[1]);
   *H = _mm512_reduce_add_pd(tt[2]);
-#else
-/* Hand-Code Vector processing for Intel Ivy-Bridge (AVX) */
-  double complex const* e;
-
-  int ix, iy, iz, n;
-
-#ifdef ARTED_STENCIL_LOOP_BLOCKING
-  int bx, by;
-#endif
-
-  const __m256i CONJ = _mm256_set_epi64x(1LL << 63, 0, 1LL << 63, 0);
-
-  __m256d tx = _mm256_setzero_pd();
-  __m256d ty = _mm256_setzero_pd();
-  __m256d tz = _mm256_setzero_pd();
-
-#pragma novector
-#ifdef ARTED_STENCIL_LOOP_BLOCKING
-  for(bx = 0 ; bx < NLx ; bx += BX)
-  for(by = 0 ; by < NLy ; by += BY)
-  for(ix = bx ; ix < MIN(bx+BX,NLx) ; ++ix) {
-  for(iy = by ; iy < MIN(by+BY,NLy) ; ++iy)
-#else
-  for(ix = 0 ; ix < NLx ; ++ix) {
-  for(iy = 0 ; iy < NLy ; ++iy)
-#endif
-  {
-    e = &E[ix*NLy*NLz + iy*NLz];
-
-    for(iz = 0 ; iz < NLz ; iz += 2)
-    {
-#define STENCIL_CALC(MM,PP,CC) \
-      v2 = _mm256_mul_pd(_mm256_broadcast_sd(CC),  _mm256_sub_pd(PP, MM)); \
-      v1 = _mm256_add_pd(v1, v2);
-
-#define STENCIL(IDM,IDP,CC) \
-      m = _mm256_load_pd((double *) (e + iz - IDM));     \
-      p = _mm256_load_pd((double *) (e + iz - IDP));     \
-      STENCIL_CALC(m,p,CC) \
-
-      __m256d m, p;
-      __m256d v1, v2, v3, v4;
-
-      // conj(e[iz])
-      __m256d ez = _mm256_load_pd((double *)(e + iz));
-      __m256d w = (__m256d) _mm256_xor_si256((__m256i) ez, CONJ);
-
-      /* x-dimension (NLy*NLz stride)  */
-      {
-        v1 = _mm256_setzero_pd();
-        STENCIL(IDX(-1), IDX(1), C+0);
-        STENCIL(IDX(-2), IDX(2), C+1);
-        STENCIL(IDX(-3), IDX(3), C+2);
-        STENCIL(IDX(-4), IDX(4), C+3);
-        v3 = _mm256_shuffle_pd(v1, v1, 0x05);
-        v4 = _mm256_mul_pd(w, v3);
-        tx = _mm256_add_pd(v4, tx);
-      }
-
-      /* y-dimension (NLz stride) */
-      {
-        v1 = _mm256_setzero_pd();
-        STENCIL(IDY(-1), IDY(1), C+4);
-        STENCIL(IDY(-2), IDY(2), C+5);
-        STENCIL(IDY(-3), IDY(3), C+6);
-        STENCIL(IDY(-4), IDY(4), C+7);
-        v3 = _mm256_shuffle_pd(v1, v1, 0x05);
-        v4 = _mm256_mul_pd(w, v3);
-        ty = _mm256_add_pd(v4, ty);
-      }
-
-      /* z-dimension (unit stride) */
-      {
-        __m256d z0,z1,z2,z3,z4,z5,z6,z7;
-#ifdef ARTED_DOMAIN_POWER_OF_TWO
-        z1 = _mm256_load_pd((double *)(e + ((iz - 2 + NLz) & (NLz - 1))));
-        z2 = _mm256_load_pd((double *)(e + ((iz + 2 + NLz) & (NLz - 1))));
-        z0 = _mm256_load_pd((double *)(e + ((iz - 4 + NLz) & (NLz - 1))));
-        z3 = _mm256_load_pd((double *)(e + ((iz + 4 + NLz) & (NLz - 1))));
-#else
-        z1 = _mm256_load_pd((double *)(e + modz[iz - 2 + NLz]));
-        z2 = _mm256_load_pd((double *)(e + modz[iz + 2 + NLz]));
-        z0 = _mm256_load_pd((double *)(e + modz[iz - 4 + NLz]));
-        z3 = _mm256_load_pd((double *)(e + modz[iz + 4 + NLz]));
-#endif
-        z6 = _mm256_permute2f128_pd(z0, z1, 0x21);
-        z4 = _mm256_permute2f128_pd(z1, ez, 0x21);
-        z5 = _mm256_permute2f128_pd(ez, z2, 0x21);
-        z7 = _mm256_permute2f128_pd(z2, z3, 0x21);
-
-        v1 = _mm256_setzero_pd();
-        STENCIL_CALC(z1, z2, C+ 9);
-        STENCIL_CALC(z0, z3, C+11);
-        STENCIL_CALC(z4, z5, C+ 8);
-        STENCIL_CALC(z6, z7, C+10);
-        v3 = _mm256_shuffle_pd(v1, v1, 0x05);
-        v4 = _mm256_mul_pd(w, v3);
-        tz = _mm256_add_pd(v4, tz);
-      }
-    }  /* NLz */
-  } /* NLy */
-  } /* NLx */
-
-  __declspec(align(64)) double r[4];
-
-  _mm256_store_pd(r, tx);
-  *F = r[0] + r[1] + r[2] + r[3];
-
-  _mm256_store_pd(r, ty);
-  *G = r[0] + r[1] + r[2] + r[3];
-
-  _mm256_store_pd(r, tz);
-  *H = r[0] + r[1] + r[2] + r[3];
-#endif /* if defined(__KNC__) || defined(__AVX512F__) */
 }
