+++ stencil/F90/hpsi_original.f90
