+++ stencil/F90/total_energy.f90
